Added functionality to manually control state of RTS and DTR lines when port has been opened with no flow control or software flow control enabled.
Added two QCheckBoxes to the control panel that are using to manually change state of RTS and DTR lines.
Implemented visibility setting of these widgets depending on selected type of flow control. When hardware flow control is enabled then they are invisible.
RTS and DTR checkboxes are enabled only when port is open.
When serial port is being opened with software flow control enabled then RTS/DTR are disabled (sets to logic low) as was so far when opening port with no flow control.