Implements popup widget with buttons which represent control characters in range 0x00-0x1F.
This widget is constructed to act as popup widget and has custom frame. On hover over buttons on the right side is displaying value of control character in decimal and hex format.
Supports inserting in hex mode.
Widget is positioning according to width of input field, length of text inside input field and current cursor position.